move image types
