+++ tsconfig.json
@@ -11,6 +11,7 @@
       "@/app/*": ["app/*"],
       "@/components/*": ["app/components/*"],
       "@/api/*": ["app/api/*"],
+      "@/types/*": ["types/*"],
       "@/lib/*": ["lib/*"],
       "@/styles/*": ["styles/*"]
     },
